test out multi-scale algorithm
The current autocorrelation computation is O(n^2), which can be slow for
large windows. Since this is a tuner with a fixed set of frequencies to
process, we can probably optimize by first applying the algorithm to
down-sampled inputs. This approach would first identify the correct
frequency range efficiently. A higher-resolution frequency search can
then be performed on a much smaller sample, significantly improving
performance. Let's try